[CHORE] 자잘한 코드 수정

+++ thanks-diary/Thanks Diary/Util/Constant/Enum/EmailLoginErrorList.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-enum EmailErrorList: NSInteger {
+enum EmailErrorList: Int {
     case ALREADY_SIGNUP = 17007
     case WRONG_EMAIL_FORMAT = 17008
     case MISMATCH_PASSWORD = 17009
